requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
pip==20.2.4
requests==2.24.0
setuptools==50.3.2
urllib3==1.25.10
python-dotenv==0.14.0
simplejson==3.17.2
numpy==1.19.2
pandas==1.1.3
python-dateutil==2.8.1
pytz==2020.1
